+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.0] - 2025-12-02
+
+### Added
+
+- OpenAPI spec import.
+- File chooser filters by suffix.
+
 ## [0.8.0] - 2025-11-20
 
 ### Added